+++ praw/models/__init__.py
@@ -19,6 +19,7 @@ from .reddit.modmail import ModmailAction, ModmailConversation, ModmailMessage
 from .reddit.more import MoreComments
 from .reddit.multi import Multireddit
 from .reddit.redditor import Redditor
+from .reddit.removal_reasons import RemovalReason
 from .reddit.submission import Submission
 from .reddit.subreddit import Subreddit
 from .reddit.widgets import (