+++ praw/models/reddit/mixins/__init__.py
@@ -18,6 +18,32 @@ class ThingModerationMixin:
 
     REMOVAL_MESSAGE_API = None
 
+    def _add_removal_reason(self, mod_note="", reason_id=None):
+        """Add a removal reason for a Comment or Submission.
+
+        :param mod_note: A message for the other moderators.
+        :param reason_id: The removal reason ID.
+
+        It is necessary to first call :meth:`~.remove` on the
+        :class:`~.Comment` or :class:`~.Submission`.
+
+        If ``reason_id`` is not specified, ``mod_note`` cannot be blank.
+
+        """
+        if not reason_id and not mod_note:
+            raise ValueError(
+                "mod_note cannot be blank if reason_id is not specified"
+            )
+        # Only the first element of the item_id list is used.
+        data = {
+            "item_ids": [self.thing.fullname],
+            "mod_note": mod_note,
+            "reason_id": reason_id,
+        }
+        self.thing._reddit.post(
+            API_PATH["removal_reasons"], data={"json": dumps(data)}
+        )
+
     def approve(self):
         """Approve a :class:`~.Comment` or :class:`~.Submission`.
 
@@ -48,9 +74,9 @@ class ThingModerationMixin:
         :param how: One of 'yes', 'no', 'admin', 'special'. 'yes' adds a
             moderator level distinguish. 'no' removes any distinction. 'admin'
             and 'special' require special user privileges to use.
-        :param sticky: Comment is stickied if True, placing it at the top of
-            the comment page regardless of score. If thing is not a top-level
-            comment, this parameter is silently ignored.
+        :param sticky: Comment is stickied if ``True``, placing it at the top
+            of the comment page regardless of score. If thing is not a
+            top-level comment, this parameter is silently ignored.
 
         Example usage:
 
@@ -72,7 +98,7 @@ class ThingModerationMixin:
         self.thing._reddit.post(API_PATH["distinguish"], data=data)
 
     def ignore_reports(self):
-        """Ignore future reports on a Comment or Submission.
+        """Ignore future reports on a :class:`~.Comment` or :class:`~.Submission`.
 
         Calling this method will prevent future reports on this Comment or
         Submission from both triggering notifications and appearing in the
@@ -86,7 +112,7 @@ class ThingModerationMixin:
            # ignore future reports on a comment:
            comment = reddit.comment('dkk4qjd')
            comment.mod.ignore_reports()
-           # ignore future reports on a submission
+           # ignore future reports on a submission:
            submission = reddit.submission(id='5or86n')
            submission.mod.ignore_reports()
 
@@ -98,7 +124,7 @@ class ThingModerationMixin:
         )
 
     def lock(self):
-        """Lock the a :class:`~.Comment` or :class:`~.Submission`.
+        """Lock a :class:`~.Comment` or :class:`~.Submission`.
 
         Example usage:
 
@@ -118,11 +144,16 @@ class ThingModerationMixin:
             API_PATH["lock"], data={"id": self.thing.fullname}
         )
 
-    def remove(self, spam=False):
+    def remove(self, spam=False, mod_note="", reason_id=None):
         """Remove a :class:`~.Comment` or :class:`~.Submission`.
 
+        :param mod_note: A message for the other moderators.
         :param spam: When True, use the removal to help train the Subreddit's
             spam filter (default: False).
+        :param reason_id: The removal reason ID.
+
+        If either ``reason_id`` or ``mod_note`` are provided, a second API
+        call is made to add the removal reason.
 
         Example usage:
 
@@ -134,10 +165,16 @@ class ThingModerationMixin:
            # remove a submission
            submission = reddit.submission(id='5or86n')
            submission.mod.remove()
+           # remove a submission with a removal reason
+           reason = reddit.subreddit.mod.removal_reasons["110ni21zo23ql"]
+           submission = reddit.submission(id="5or86n")
+           submission.mod.remove(reason_id=reason.id)
 
         """
         data = {"id": self.thing.fullname, "spam": bool(spam)}
         self.thing._reddit.post(API_PATH["remove"], data=data)
+        if any([reason_id, mod_note]):
+            self._add_removal_reason(mod_note, reason_id)
 
     def send_removal_message(
         self,
@@ -145,7 +182,7 @@ class ThingModerationMixin:
         title="ignored",
         type="public",  # pylint: disable=redefined-builtin
     ):
-        """Send a removal message for a Comment or Submission.
+        """Send a removal message for a :class:`~.Comment` or :class:`~.Submission`.
 
         Reddit adds human-readable information about the object to the message.
 
@@ -176,7 +213,7 @@ class ThingModerationMixin:
         return self.thing._reddit.post(url, data={"json": dumps(data)}) or None
 
     def undistinguish(self):
-        """Remove mod, admin, or special distinguishing on object.
+        """Remove mod, admin, or special distinguishing from an object.
 
         Also unstickies the object if applicable.
 
@@ -199,8 +236,9 @@ class ThingModerationMixin:
     def unignore_reports(self):
         """Resume receiving future reports on a Comment or Submission.
 
-        Future reports on this Comment or Submission will cause notifications,
-        and appear in the various moderation listings.
+        Future reports on this :class:`~.Comment` or :class:`~.Submission`
+        will cause notifications, and appear in the various moderation
+        listings.
 
         Example usage:
 
@@ -209,7 +247,7 @@ class ThingModerationMixin:
            # accept future reports on a comment:
            comment = reddit.comment('dkk4qjd')
            comment.mod.unignore_reports()
-           # accept future reports on a submission
+           # accept future reports on a submission:
            submission = reddit.submission(id='5or86n')
            submission.mod.unignore_reports()
 
@@ -221,7 +259,7 @@ class ThingModerationMixin:
         )
 
     def unlock(self):
-        """Unlock the a :class:`~.Comment` or :class:`~.Submission`.
+        """Unlock a :class:`~.Comment` or :class:`~.Submission`.
 
         Example usage: