+++ praw/models/user.py
@@ -100,10 +100,26 @@ class User(PRAWBase):
         return self._me
 
     def moderator_subreddits(self, **generator_kwargs):
-        """Return a ListingGenerator of subreddits the user is a moderator of.
+        """(DEPRECATED) Return a ListingGenerator of subreddits the user moderates.
+
+        ..warning:: (Deprecated) This method will be removed in the next major
+                    version of PRAW. Please use :meth:`.Redditor.moderated`
+                    instead.
 
         Additional keyword arguments are passed in the initialization of
         :class:`.ListingGenerator`.
+
+        .. note:: This method will return a maximum of 100 moderated
+           subreddits, ordered by subscriber count. To retrieve more than
+           100 moderated subreddits, please see :meth:`.Redditor.moderated`.
+
+        Usage:
+
+        .. code-block:: python
+
+           for subreddit in reddit.user.moderator_subreddits():
+               print(subreddit.display_name)
+
 
         """
         return ListingGenerator(