+++ tests/integration/models/reddit/test_submission.py
@@ -360,6 +360,50 @@ class TestSubmissionModeration(IntegrationTest):
             self.reddit.submission("4b536h").mod.remove(spam=True)
 
     @mock.patch("time.sleep", return_value=None)
+    def test_remove_with_reason_id(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubmissionModeration.test_remove_with_reason_id"
+        ):
+            self.reddit.submission("e3op46").mod.remove(
+                reason_id="110nhral8vygf"
+            )
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_add_removal_reason(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubmissionModeration.test_add_removal_reason"
+        ):
+            submission = Submission(self.reddit, "e3oo6a")
+            submission.mod.remove()
+            submission.mod._add_removal_reason(
+                mod_note="Foobar", reason_id="110nhral8vygf"
+            )
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_add_removal_reason_without_id(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubmissionModeration.test_add_removal_reason_without_id"
+        ):
+            submission = Submission(self.reddit, "e3om6k")
+            submission.mod.remove()
+            submission.mod._add_removal_reason(mod_note="Foobar")
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_add_removal_reason_without_id_or_note(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubmissionModeration.test_add_removal_reason_invalid"
+        ):
+            with pytest.raises(ValueError) as excinfo:
+                submission = Submission(self.reddit, "e4ds11")
+                submission.mod.remove()
+                submission.mod._add_removal_reason()
+            assert excinfo.value.args[0].startswith("mod_note cannot be blank")
+
+    @mock.patch("time.sleep", return_value=None)
     def test_send_removal_message(self, _):
         self.reddit.read_only = False
         with self.recorder.use_cassette(